Add utility functions
We may need some that will be DRY. Stating it in our utils file, may be
the best way of sharing it across the library.

In the current state, we're dealing with the response body taken from
the API and assigning it to the struct for further use withing golang.